% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{episode_data}
\alias{episode_data}
\title{Bob's Burgers Episode Data}
\format{
A data frame with 309 rows and 14 variables:
\describe{
\item{episode_overall}{The overall episode number in the series (1-309).}
\item{season}{The season number (1-16).}
\item{episode}{The episode number within the season.}
\item{title}{The title of the episode.}
\item{aired_date}{The date the episode originally aired (Date format).}
\item{year}{The year the episode aired.}
\item{rating}{The TMDB community rating of the episode (scale 1-10). Note: Prior to v0.2.0, ratings were sourced from IMDb. TMDB and IMDb ratings are from different user communities and are not directly comparable.}
\item{votes}{The number of TMDB user votes for the episode rating.}
\item{synopsis}{A brief description of the episode's plot from TMDB.}
\item{directed_by}{The name(s) of the director(s), from Wikipedia.}
\item{written_by}{The name(s) of the writer(s), from Wikipedia.}
\item{us_viewers_millions}{The number of US viewers (in millions) who watched the episode when it first aired, from Wikipedia. NA for episodes where viewership data is not available.}
\item{runtime}{Episode runtime in minutes.}
\item{tmdb_id}{The unique TMDB episode identifier.}
}
}
\source{
\itemize{
\item TMDB API: \url{https://www.themoviedb.org/tv/32726-bob-s-burgers}

\item Wikipedia: \url{https://en.wikipedia.org/wiki/List_of_Bob\%27s_Burgers_episodes}
}
}
\usage{
episode_data
}
\description{
A dataset containing information about Bob's Burgers episodes, including
ratings, air dates, synopses, and production credits. Data is compiled from
TMDB (The Movie Database) and Wikipedia.
}
\details{
This dataset was renamed from \code{imdb_wikipedia_data} in v0.2.0. The rating
source changed from IMDb to TMDB because IMDb implemented bot protection that
prevents web scraping. TMDB provides an official API with complete episode data.
}
\examples{
# Load the data
data(episode_data)

# View structure
str(episode_data)

# Summary statistics
summary(episode_data$rating)

# Count episodes per season
table(episode_data$season)

}
\keyword{datasets}
